==> notebooks/step3_portfolio_selection.ipynb <==
### Diversification and Top Performers by Sector

1. **Notebook Objective:**
	- This notebook will group stocks by sector
	- It will identify the top *n* stocks in each sector based on user-defined criteria (n is defined in `config.py`)
	- It will compute the correlation matrix of returns across selected stocks
		- Notation: Let $R$ be a matrix of returns, then the correlation matrix $\rho$ is given by  
		  $\rho_{i,j} = \frac{\text{Cov}(R_i, R_j)}{\sigma_{R_i} \cdot \sigma_{R_j}}$
	- It will identify the most diversifiable stocks (those with the lowest correlations)
	- Finally, it will export potential portfolios—those with high Sharpe Ratios and strong performance metrics—to `config.py` for use in later analysis

2. **Bugs**
    - Need to fix function get_corr_pairs in utils.finance_utils.py to allow user to match date with program


import time
import os 
import sys

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), '..')))
print(f'Current Working Directory: {os.getcwd()}')

from utils.finance_utils import *
from utils.helpers import separate_corr_pairs
from utils.config import TOP_N_STOCKS, PROGRAM_START_DATE, PROGRAM_END_DATE

Sector Valuation: Group by sectors, and find best stock in each sector

start_time = time.time() # measure run time

filtered_valuation_df = pd.read_csv(f'../data/clean/filtered_valuation_df.csv')

top_stocks_on_profit = get_top_n_by_sector(filtered_valuation_df, filter_var='profitMargins', top_n=3)
top_stocks_on_sharpe = get_top_n_by_sector(filtered_valuation_df, filter_var='Sharpe_ratios', top_n=3)

top_stocks_on_profit

Diversifaction: Find negative correlated stocks

top_profit_corr_pairs = get_corr_pairs_of_stocks(top_stocks_on_profit['Tickers'].tolist(), start_date=PROGRAM_START_DATE, end_date=PROGRAM_END_DATE)
top_sharpe_corr_pairs = get_corr_pairs_of_stocks(top_stocks_on_sharpe['Tickers'].tolist(), start_date=PROGRAM_START_DATE, end_date=PROGRAM_END_DATE)

mpt_stocks_for_sharpe = separate_corr_pairs(top_sharpe_corr_pairs.index.to_list(), top_n=TOP_N_STOCKS)
mpt_stocks_for_profit = separate_corr_pairs(top_profit_corr_pairs.index.to_list(), top_n=TOP_N_STOCKS)

print(f'Top Stocks based on Sharpe: {mpt_stocks_for_sharpe}')
print(f'Top Stocks based on Profit Margin: {mpt_stocks_for_profit}')

top_sharpe_corr_pairs

Seperate low corr pairs

with open('../utils/config.py', 'a') as f:
    f.write('\n# New Entry from portfolio_selection.ipynb')
    f.write(f'\nmpt_stocks_for_sharpe = {(mpt_stocks_for_sharpe).tolist()}\n')
    f.write(f'mpt_stocks_for_profit = {(mpt_stocks_for_profit).tolist()}\n')
    # f.write(f'my_pairs = {(my_pairs).tolist()}\n')

filtered_valuation_df[filtered_valuation_df['Tickers'].isin(mpt_stocks_for_sharpe)].reset_index(drop=True)

filtered_valuation_df[filtered_valuation_df['Tickers'].isin(mpt_stocks_for_profit)].reset_index(drop=True)
